athene: defconfig: Use 64bit Binder API
IPC_32BIT originally was intended to be used on 4.4 and older userspaces, because they
didn't support new API and 64-bit pointers, but then it was forced on 32-bit platforms
because ARM didn't have essential 64-bit get_user() calls. The main limitation of this
option is that it is incompatible with new API, so kernel and userspace should have same
binder version to work properly. Necessary calls were later introduced in commit "ARM:
8091/2: add get_user() support for 8 byte types" and now AOSP and mainline are going
towards removing old 32-bit binder API support. 64-bit binder is 100% compatible with
32-bit kernel and userspace,
so there is no reason for us to keep using the old solution anymore.

Change-Id: I91ce91a1a17ceb7110184657663e957bd8793d3f